community_usecase/a_share_investment_agent_camel/requirements.txt
camel-ai[all]==0.2.36
pandas==2.0.3
numpy==1.24.4
python-dotenv==1.0.0
akshare==1.11.57
matplotlib==3.7.3
pydantic==1.10.12
tqdm==4.66.1
plotly==5.14.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
scikit-learn==1.2.2
statsmodels==0.14.0
pyecharts==2.0.4
jieba==0.42.1
ta-lib==0.4.28
pytest==7.4.3
black==23.11.0
isort==5.12.0
mypy==1.7.1
google-generativeai==0.3.1
openai==1.6.1
loguru==0.7.2
pytest-cov==4.1.0 